+++ changelog.d/mobile-bottom-sheet-options.md
@@ -1,0 +1,1 @@
+- Add a mobile bottom-sheet options view with the `mobileOptionsMode` override prop.
