+++ example/screens/mobile-sheet-resize-screen.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+import MobileSheetSelectScreen from "./mobile-sheet-select-screen"
+
+export default function MobileSheetResizeScreen() {
+  return <MobileSheetSelectScreen />
+}
