+++ example/screens/multiple-clear-screen.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+import MultipleSelectScreen from "./multiple-select-screen"
+
+export default function MultipleClearScreen() {
+  return <MultipleSelectScreen />
+}
