+++ example/screens/pagination-change-page-screen.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+import PaginationSelectScreen from "./pagination-select-screen"
+
+export default function PaginationChangePageScreen() {
+  return <PaginationSelectScreen />
+}
