+++ example/screens/pagination-manual-entry-screen.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+import PaginationSelectScreen from "./pagination-select-screen"
+
+export default function PaginationManualEntryScreen() {
+  return <PaginationSelectScreen />
+}
