+++ example/screens/pagination-next-prev-screen.tsx
@@ -1,0 +1,7 @@
+import React from "react"
+
+import PaginationSelectScreen from "./pagination-select-screen"
+
+export default function PaginationNextPrevScreen() {
+  return <PaginationSelectScreen />
+}
